Moved game descriptor keys, added ALL keys to getAllActivities()
