+++ test/gb/activity.test.ts
@@ -47,7 +47,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=20-04-2021&sort=-date`,
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29%2CNutrition_Diary%2CLOG_INSULIN%2CBLOOD_GLUCOSE_MSMT&start=19-04-2021&end=20-04-2021&sort=-date`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
@@ -65,7 +65,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=21-04-2021&sort=-date`,
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29%2CNutrition_Diary%2CLOG_INSULIN%2CBLOOD_GLUCOSE_MSMT&start=19-04-2021&end=21-04-2021&sort=-date`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
@@ -89,7 +89,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=21-04-2021&sort=-date&limit=10&page=1`,
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29%2CNutrition_Diary%2CLOG_INSULIN%2CBLOOD_GLUCOSE_MSMT&start=19-04-2021&end=21-04-2021&sort=-date&limit=10&page=1`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
@@ -106,7 +106,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=20-04-2021&sort=-date`,
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29%2CNutrition_Diary%2CLOG_INSULIN%2CBLOOD_GLUCOSE_MSMT&start=19-04-2021&end=20-04-2021&sort=-date`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
@@ -126,7 +126,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=21-04-2021&sort=-date`,
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29%2CNutrition_Diary%2CLOG_INSULIN%2CBLOOD_GLUCOSE_MSMT&start=19-04-2021&end=21-04-2021&sort=-date`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
@@ -162,7 +162,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=21-04-2021&sort=-date&gds=WALK`,
+                url: `${endpoint}/players/0/activities?gds=WALK&start=19-04-2021&end=21-04-2021&sort=-date`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
@@ -181,7 +181,7 @@ describe('with mocked activities get call', () => {
         expect(request).toHaveBeenCalledTimes(1);
         expect(request).toHaveBeenCalledWith(
             expect.objectContaining({
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=20-04-2021&sort=-date&gds=WALK`,
+                url: `${endpoint}/players/0/activities?gds=WALK&start=19-04-2021&end=20-04-2021&sort=-date`,
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 })
