+++ test/gb/exercise.test.ts
@@ -3,7 +3,8 @@ import { TokenHandler } from '../../src/gb/auth/tokenHandler';
 import { GameBusClient } from '../../src/gb/gbClient';
 import { ExerciseModel } from '../../src/gb/models/exerciseModel';
 import { ActivityGETData } from '../../src/gb/models/gamebusModel';
-import { Exercise, ExerciseGameDescriptorNames } from '../../src/gb/objects/exercise';
+import { Exercise } from '../../src/gb/objects/exercise';
+import { ExerciseGameDescriptorNames } from '../../src/gb/objects/keys';
 import { mockRequest } from '../testUtils/requestUtils';
 
 jest.mock('axios');
@@ -77,7 +78,7 @@ describe('with mocked exercises get call', () => {
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 }),
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=21-04-2021&sort=-date&gds=WALK`
+                url: `${endpoint}/players/0/activities?gds=WALK&start=19-04-2021&end=21-04-2021&sort=-date`
             })
         );
         expect(exercises).toEqual([]);
@@ -96,7 +97,7 @@ describe('with mocked exercises get call', () => {
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 }),
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=21-04-2021&sort=-date&gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29`
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29&start=19-04-2021&end=21-04-2021&sort=-date`
             })
         );
         expect(exercises).toEqual([]);
@@ -118,7 +119,7 @@ describe('with mocked exercises get call', () => {
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 }),
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=20-04-2021&sort=-date&gds=WALK`
+                url: `${endpoint}/players/0/activities?gds=WALK&start=19-04-2021&end=20-04-2021&sort=-date`
             })
         );
         expect(exercises).toEqual([]);
@@ -136,7 +137,7 @@ describe('with mocked exercises get call', () => {
                 headers: expect.objectContaining({
                     Authorization: `Bearer ${mockToken}`
                 }),
-                url: `${endpoint}/players/0/activities?start=19-04-2021&end=20-04-2021&sort=-date&gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29`
+                url: `${endpoint}/players/0/activities?gds=WALK%2CRUN%2CBIKE%2CSOCCER%2CBASKETBALL%2CVOLLEYBALL%2CRUGBY%2CBASEBALL%2CHORSE_RIDING%2CATHLETICS%2CSWIMMING%2CWATER_POLO%2CSURFING%2CGOLF%2CLACROSSE%2CTENNIS%2CSQUASH%2CBADMINTON%2CTABLE_TENNIS%2CSKIING%2CICE_HOCKEY%2CFIELD_HOCKEY%2CICE_SKATING%2CROLLER_SKATING%2CFITNESS%2CYOGA%2CAEROBICS%2CMARTIAL_ARTS%2CDANCE%2CPOOL%2CDARTS%2CAIR_HOCKEY%2CBOWLING%2CCHESS%2CGYMNASTICS%2CHIKE%2CMOUNTAINBIKE%2CWALK%28DETAIL%29%2CRUN%28DETAIL%29%2CBIKE%28DETAIL%29&start=19-04-2021&end=20-04-2021&sort=-date`
             })
         );
         expect(exercises).toEqual([]);
